🎨 feat: add 'link' class to technology and hobby links for improved sty… (#170)
This pull request makes a small styling update to the
`createTechStackSection` and `createHobbiesSection` functions in
`assets/js/content.js`. The main change is the addition of the `link`
class to certain anchor elements, likely to improve or standardize their
appearance.

- UI Consistency:
* Added the `link` CSS class to anchor tags generated for technologies
and hobbies, ensuring consistent styling for these links in both the
tech stack and hobbies sections.
[[1]](diffhunk://#diff-2de343f41e8c9bc4a79c58c52c6170765675f8b437643ca73b7813911a23b97eL211-R211)
[[2]](diffhunk://#diff-2de343f41e8c9bc4a79c58c52c6170765675f8b437643ca73b7813911a23b97eL287-R287)